test utility functions moved to separate folder, instruction for running added
