+++ README.md
@@ -6,10 +6,10 @@ Dark pool on Miden: A privacy-preserving dark pool trading system built on Miden
 ## Overview
 This prototype is experimenting with:
 
-	•	Private orders via MidenNotes
-	•	Off-chain matching using a TCP-based matcher(TEE to be added)
-	•	scripts for partially fillable swaps (SWAPp)
-	•	Basic CLI for submitting/cancelling orders
+	•	Private orders via MidenNotes.
+	•	Off-chain matching using a TCP-based matcher(TEE to be added).
+	•	scripts for partially fillable swaps (SWAPp). The swap note script is not in its final version and some changes needs to be made.
+	•	Basic CLI for submitting/cancelling orders.
 
 ## Current Status
 
@@ -26,7 +26,7 @@ This prototype is experimenting with:
 ❌ Secure transport (planned)
 
 
-## Building and Testing
+## Building, Testing and Running
 
 - To Build: `cargo build --release`
 
@@ -34,4 +34,34 @@ This prototype is experimenting with:
     - `cargo test --release --test user_flow -- test_open_order --exact`
     - `cargo test --release --test user_flow -- test_cancel_order --exact`
 
-Note: in case of failure of test, delete the keystore and store
+- To Run:
+    - matcher: `cargo run --release --bin matcher`
+    - user:
+        - open-order: 
+            ```sh
+            cargo run --release \
+            --bin user \
+            -- open-order \
+            --user-id <USER_ID_HEX_STRING> \
+            --offered-asset <OFFERED_ASSET_HEX_ID> \
+            --offered-amout <AMOUNT> \
+            --requested-asset <REQUESTED_ASSET_HEX_ID> \
+            --price <PRICE>
+            ```
+        - cancel-order:
+            ```sh
+            cargo run --release \
+            --bin user \
+            -- cancel-order \
+            --user-id <USER_ID_HEX_STRING> \
+            --order-id <SWAP_NOTE_HEX_ID> \
+            --tag <SWAP_NOTE_TAG>
+            ```
+        - consume-swapped:
+            ```sh
+            cargo run --release \
+            --bin user \
+            -- consume-swapped \
+            --user-id <USER_ID_HEX_STRING>
+            ```
+Note(Only in case of testing): in case of failure of test, delete the keystore and store
