+++ src/utils/common.rs
@@ -114,34 +114,6 @@ pub fn create_partial_swap_note(
     Ok(note)
 }
 
-pub async fn delete_keystore_and_store() {
-    // Remove the SQLite store file
-
-    let keystore_dir: &str = &format!("./keystore");
-    let store_path: &str = &format!("./store.sqlite3");
-
-    if tokio::fs::metadata(store_path).await.is_ok() {
-        if let Err(e) = tokio::fs::remove_file(store_path).await {
-            eprintln!("failed to remove {}: {}", store_path, e);
-        }
-    } else {
-        println!("store not found: {}", store_path);
-    }
-
-    // Remove all files in the ./keystore directory
-    match tokio::fs::read_dir(keystore_dir).await {
-        Ok(mut dir) => {
-            while let Ok(Some(entry)) = dir.next_entry().await {
-                let file_path = entry.path();
-                if let Err(e) = tokio::fs::remove_file(&file_path).await {
-                    eprintln!("failed to remove {}: {}", file_path.display(), e);
-                }
-            }
-        }
-        Err(e) => eprintln!("failed to read directory {}: {}", keystore_dir, e),
-    }
-}
-
 /// Generates a SWAP note tag
 /// build_swap_tag(note_type, asset1, asset2)
 /// where asset_{i} is an Asset created with AssetId of the asset pairs and 0 amount so that the tag is deterministic for a given asset pair
