+++ tests/utils/test_utils.rs
@@ -1,36 +1,30 @@
 use miden_client::Client;
+use miden_client::account::Account;
+use miden_client::keystore::FilesystemKeyStore;
+use rand::rngs::StdRng;
+
 use miden_client::ClientError;
 use miden_client::account::{
-    Account, AccountBuilder, AccountStorageMode, AccountType, component::BasicFungibleFaucet,
+    AccountBuilder, AccountStorageMode, AccountType, component::BasicFungibleFaucet,
     component::BasicWallet, component::RpoFalcon512,
 };
 
-#[allow(unused_imports)]
-use miden_objects::note::{
-    Note, NoteAssets, NoteExecutionHint, NoteExecutionMode, NoteInputs, NoteMetadata,
-    NoteRecipient, NoteScript, NoteTag, NoteType,
-};
+use miden_objects::note::NoteType;
+use miden_objects::{Felt, asset::FungibleAsset};
 
-#[allow(unused_imports)]
-use miden_objects::{
-    Felt, NoteError, Word,
-    account::AccountId,
-    asset::{Asset, FungibleAsset},
-};
-
-#[allow(unused_imports)]
 use miden_client::{
-    asset::TokenSymbol,
-    auth::AuthSecretKey,
-    builder::ClientBuilder,
-    crypto::SecretKey,
-    keystore::FilesystemKeyStore,
-    rpc::{Endpoint, TonicRpcClient},
+    asset::TokenSymbol, auth::AuthSecretKey, crypto::SecretKey,
     transaction::TransactionRequestBuilder,
 };
 
 use rand::RngCore;
 use std::time::Duration;
+
+#[derive(Debug, Clone)]
+pub struct TestUser {
+    pub user_id: String,
+    pub account_id: Account,
+}
 
 pub async fn create_account(
     client: &mut Client,
@@ -163,4 +157,118 @@ pub async fn mint_and_consume(
 
     client.sync_state().await?;
     Ok(())
+}
+
+pub async fn setup_test_user(
+    mut client: &mut Client,
+    keystore: FilesystemKeyStore<StdRng>,
+    user_id: &str,
+    faucet_a: Account,
+    faucet_b: Account,
+    amount: u64,
+) -> TestUser {
+    let sync_summary = client.sync_state().await.unwrap();
+    println!("Latest block: {}", sync_summary.block_num);
+
+    let account = create_account(&mut client, keystore.clone()).await.unwrap();
+
+    client.sync_state().await.unwrap();
+    // Mint token A to the user
+    mint_and_consume(&mut client, faucet_a, account.clone(), amount)
+        .await
+        .unwrap();
+
+    mint_and_consume(&mut client, faucet_b, account.clone(), 20)
+        .await
+        .unwrap();
+
+    let _ = client.sync_state().await.unwrap();
+
+    TestUser {
+        user_id: user_id.to_string(),
+        account_id: account,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use miden_client::builder::ClientBuilder;
+    use miden_client::rpc::Endpoint;
+    use miden_client::rpc::TonicRpcClient;
+    use std::sync::Arc;
+
+    use super::*;
+    #[tokio::test]
+    async fn test_setup() {
+        let endpoint = Endpoint::new(
+            "https".to_string(),
+            "rpc.testnet.miden.io".to_string(),
+            Some(443),
+        );
+        let timeout_ms = 10_000;
+        let rpc_api = Arc::new(TonicRpcClient::new(&endpoint, timeout_ms));
+
+        let mut client = ClientBuilder::new()
+            .with_rpc(rpc_api)
+            .with_filesystem_keystore("./keystore")
+            .in_debug_mode(true)
+            .build()
+            .await
+            .unwrap();
+
+        let sync_summary = client.sync_state().await.unwrap();
+        println!("Latest block: {}", sync_summary.block_num);
+
+        let keystore = FilesystemKeyStore::new("./keystore".into()).unwrap();
+        let symbol = "ETH";
+
+        let faucet_a = create_faucet(&mut client, keystore.clone(), symbol)
+            .await
+            .unwrap();
+
+        let faucet_b = create_faucet(&mut client, keystore.clone(), symbol)
+            .await
+            .unwrap();
+
+        let mut users: Vec<TestUser> = Vec::new();
+
+        let user = setup_test_user(
+            &mut client,
+            keystore.clone(),
+            &format!("testuser"),
+            faucet_a.clone(),
+            faucet_b.clone(),
+            100,
+        )
+        .await;
+        users.push(user);
+    }
+}
+
+pub async fn delete_keystore_and_store() {
+    // Remove the SQLite store file
+
+    let keystore_dir: &str = &format!("./keystore");
+    let store_path: &str = &format!("./store.sqlite3");
+
+    if tokio::fs::metadata(store_path).await.is_ok() {
+        if let Err(e) = tokio::fs::remove_file(store_path).await {
+            eprintln!("failed to remove {}: {}", store_path, e);
+        }
+    } else {
+        println!("store not found: {}", store_path);
+    }
+
+    // Remove all files in the ./keystore directory
+    match tokio::fs::read_dir(keystore_dir).await {
+        Ok(mut dir) => {
+            while let Ok(Some(entry)) = dir.next_entry().await {
+                let file_path = entry.path();
+                if let Err(e) = tokio::fs::remove_file(&file_path).await {
+                    eprintln!("failed to remove {}: {}", file_path.display(), e);
+                }
+            }
+        }
+        Err(e) => eprintln!("failed to read directory {}: {}", keystore_dir, e),
+    }
 }
